[REFACT] 추천 대외활동 리스트 분리하기 #NADA-43
